feat: Add answer_hello function to handle incoming messages
This commit adds the `answer_hello` function to handle incoming messages in the flet-chat messaging app. The function takes a `Message` object as input and returns a new `Message` object with a greeting for the user. This function will be used to generate a response when a user sends a "hello" message in the chat.